+++ task_app/api/views.py
@@ -40,6 +40,7 @@ class TaskDetailView(APIView):
         
         def patch(self, request, pk):
             tasks = CreateTask.objects.get(pk=pk)
+            self.check_object_permissions(request, tasks)
             serializer = TaskSerializer(tasks, data=request.data, partial=True)
             if serializer.is_valid():
                 serializer.save()
@@ -48,6 +49,7 @@ class TaskDetailView(APIView):
         
         def delete(self, request, pk):
                 tasks = CreateTask.objects.get(pk=pk)
+                self.check_object_permissions(request, tasks)
                 tasks.delete()
                 return Response({"message" : "successfully deleted",}, status=204)
